Fix for bug 67 regression test
This is very similar to my previous commit, for the invalidMath test.

Both of them are error message tests.  And, both of them were errors
that the standard flex/bison error trapping was either missing, or
having trouble with.

As far as I can tell, when such cases emerge, the only way to deal with
them is to create new rules in the parser that explicitly match that
error condition.    I'm not crazy about having to do that, as it implies
that a near infinite number of rules could someday be required to catch
various errors.

Note, MOST of the errors in the regression suite are
caught just fine by the parser, without special rules.   But this is a
pattern I'd rather not repeat more than necessary.

This brings the number of failing tests down to 12 (excluding failures
related to Xyce branch divergence).