Pass Start/Stop Sweep Vals into AC/NOISE Measure Objects
This commit changes the function signatures of the updateAC() and
updateNoise() functions of the measure classes to pass in the
start/stop frequencies specified on the .AC or .NOISE lines. (Note:
This approach is consistent with how the AC outputters work.)  This
additional information will be useful for the processing the AVG,
DERIV-WHEN, FIND-WHEN, INTEG, RMS and WHEN measures, as I work on
making the FROM-TO and RISE/FALL/CROSS qualifiers work better for
those measure types.

This commit has no immediate impact on the behavior of those
measure types though.

SON Bugs 1304, 1309 and 1310
http://joseki-vm.sandia.gov/bugzilla/show_bug.cgi?id=1304
http://joseki-vm.sandia.gov/bugzilla/show_bug.cgi?id=1309
http://joseki-vm.sandia.gov/bugzilla/show_bug.cgi?id=1310