Fix for ltra_step regression test
The ltra_step test has been failing recently with the new expression
library.  It is a relatively new failure, which I think started when I
modified some of the library's internal book keeping a few weeks ago.

This commit fixes the issue, albeit with a (probably) less than perfect
fix.

The way the new expression library works, in addition to the AST tree,
which is created by the library (mostly) during parsing, there are a lot
of "book keeping" STL containers, which are mostly used when the library
interacts with the rest of Xyce.  For example, under some circumstances
you need to get a list of all the  unique parameters used by the
expression.  Or, you need to update all the voltage nodes used by the
expresion, etc.

One way to do this could have been to just traverse the tree looking for
these things every time you need them.  I decided that this would be
inefficient, so I attempted to store all of this information in objects
such as std::vector and std::unordered_map.  maps are preferred most of
the time, to get the fastest searches.

This mostly has worked.  However, there is one glitch.  For parameters,
I created both a map and vectors of objects.  The vector is mainly there
b/c during parsing, I need to preserve the order of the parameters,
particularly when dealing with .func declarations.    And, I intended
the map to be what was used once parsing was over.  But, I didn't hold
to that, and two functions:  newExpression::make_const and
newExpression::make_var are instead using the vector.

The vector is not the best choice for those functions, but it can work
fine if it is up-to-date.  Unfortunately, b/c I believed that I had made
the decision to ONLY use the param vector during parsing, I didn't
updated it when external nodes were attached. (which happens when either
a .func or a .param are resolved).  So, the vector used by these
functions didn't contain enough information, if they were called after a
new external dependency was attached.

The best fix is to change these functions to use the maps.  That takes a
smidgen more coding than one I did, which is the second best fix.  For
the time being, I'm updating the vector after each external node
attachment.  That isn't quite the best fix, but it is a fix.