Cleanup, plus initial work on managing .func state
Two tests that still fail with the new expression library are the
"bridgewire" tests.

These tests use expressions pretty heavily, so it isn't super suprising
that they have caused problems.

The specific problem with these tests is that they use the time
integration operator, sdt, and they use this operator inside of .funcs.

A while ago, I worked on the sdt operator and got it working, except for
the .func use case.  The reason it doesn't work for the .func use case
is that my original implementation of .func assumed that all operators
were 100% dependent on their arguments, and nothing else.  But, this
isn't true for sdt and a few others.  Operators like SDT have an
internal state that must be unique for each call to that function.  So,
just like it is necessary to pass in the function arguments, it is also
necessary to pass in a state object.  I failed to set up the
infastructure for passing in a state object.

I put it aside last month because I wasn't sure how to do it.  But I've
had time to think about it and now I think I know what to do.

In this commit, the main thing that has been added (besides general
cleanup) is a new virtual function to all the ops, called "getStateOps".
This function is very similar to "getInterestingOps", except that it is
much more targeted.  Whereas "getInterestingOps" traverses the AST tree
and gathers up every kind of operator, this function traverses the tree
and only gathers operators known to require persistent internal state
to work correctly.

Also, getInterestingOps is called from functions in the newExpression
class.  This new function is intended to primarily be called directly
from the funcOp class, on the sub-tree that is attached to that class.

Once the funcOp class has gathered a container of all the
state-dependent operators that sit underneath it, it can then manage the
state information for each one.