Improvements to SDT to work with .FUNC
The bridgewire tests have been failing with the new expression library.

I believe the main reason for these failures has been that the SDT
operator is used under several layers of .FUNCs, and in the new
expression library the correct "state" for the SDT operator was getting
mangled.

This is another issue related to new expression NOT doing string
substitution to resolve .FUNC.  Since new expression instead just
attaches a node from the .FUNC, the AST sub-tree associated with that
.FUNC is identical to all the expressions that call it.

So, in my first implementation, that meant that the expression evaluated
by the .func needed to only depend upon the arguments passed into it by
the calling expression.   But, for operators like SDT, which maintain a
history, this wasn't possible.  The evaluation also depends on that
history, or the internal "state" of the operator.

For this commit, I have added infastructure to (1) in the funcOp
operator, find state-dependent AST nodes in the sub-tree and (2)
maintain a vector of "state" objects and (3) substitute these state
objects into their respective operators during evaluation. (kind of like
I already substitute the function args).

This also required modifying the SDT operator to use the state objects,
and allow them to be replaced via an accessor function.  The DDT
operator has the same problem, but I haven't fixed it in this
commit.

I also found at least one mistake in the handling of the time step (dt)
in the new expression library and fixed it as well.

I also fixed up the newExpression::clear function.    It was pretty far
out of date, and it is actually used in a few limited cases in Xyce.

Also, as usual on these commits, some cleanup is included.   I have
added more SDT unit tests to parserUnitest.C, to test the usage of SDT
inside of .funcs.

With this commit, the BRIDGEWIRE/bridgewire test now passes, but the
BRIDGEWIRE/gentleBridgewire test gets a comparison failure.  For the
test that fails, the divergence happens entirely in the latter parts of
the run, when the integral is supposed to have died of to near zero.
The new expression library gets some weird discontinuities at this
point, and then drops below zero before eventually recovering.  I spent
some time today trying to figure out why this was happening, but haven't
found the answer yet.

So, the bridgewire circuits aren't quite fixed.  Yet.