Apply state management to the ddt operator
In my previous commit, I implemented some machinery for tracking the
state of sdt operators, even when called inside of .func.

For this commit, I have done the same for ddt.  While working on this, I
also created 8 ddt unit tests.  I had not created any before now, and
doing this revealed a bug in the ddt implementation.  The bug was
causing the derivative of ddt to always be zero, which was wrong.  This
issue is fixed now.