Fix overlooked model regeneration
When I added BSIM-SOI 4.5.0 to Xyce (a rushed addition because we
found a serious bug in 4.6.1 (as delivered from BSIM group and
accepted by CMC)), I forgot to add its generation and installation as
a step in  the scripts I use to regenerate everything.  So BSIM-SOI
4.5.0 did not get generated when I fixed gitlab issue 42.

Spotted because I've been working with a user who builds Xyce on
Raspberry Pi, and was puzzled because he reports that now the build
does fairly well until it hits BSIM-SOI 4.5.0 --- that's because that
model is still the old memory pig version.

Since this cannot impact weekly testing and doesn't break the build or
nightly testing, I will be doing an almost immediate develop->master
merge so I can push this to github.