Add DIODE_CMC documentation
Add Level 2002 tables and documentation.

I have added a link to the Si2 standard models page, but I find that
the diode_cmc documentation is in fact not there, and the diode_cmc
link points to a sharepoint site that is inaccessible by random
members of the public.

It may be necessary to hunt down other links that actually provide
useful information about this CMC model, since the official CMC source
doesn't.

https://gitlab-ex.sandia.gov/Xyce/Xyce/-/46