Fix emacs xml validation issues
Emacs' XML validator sometimes has trouble when it encounters unquoted
blocks of hyphens.

Because XML (and therefore ADMST) is annoying enough to keep correct,
I do need that validator.  And for what I'm working on now, I don't
want to have to keep ignoring Emacs' highlighting of big blocks of
perfectly valid code in bright red, or the "XML:Invalid" indicator in
the status bar.

For some bizarre reason, the validation issue was only getting tripped
by one specific line of comments with hyphens, even though there were
many more of exactly the same form.  But to clear the entire mess for
now and for good, I'm changing every single such line to one that is
quoted and emitted by an "admst:text" command rather than a literal.

I have taken pains to assure that this causes NO changes in the actual
output, which means putting lots of newlines into the quoted string
that one would not think necessary.  This is because when ADMS
processes the XML it does different things with something like this:

   some literal line
   <admst:text format="foo\n"/>

and

   <admst:text format="foo\n"/>
   some literal line

In the former case, it winds up emitting:
   some literal linefoo

and in the latter case emits
   foo
   some literal line

And it gets more confusing if there are blank lines in the XML ---
sometimes they're emitted literally (generally if there's more than
one), sometimes they're just whitespace that is ignored.

The reason I took such pains is to make sure that the emitted code has
NO changes whatsoever, so I don't have to commit all the
ADMS-generated models.