Fix for sourcestepping and expressions
One failing tests with the new expression library has been the
HICUM/HICUM_L0/fgum_ac_npn_full_sh_ac test.

It didn't use expressions very much, and the expressions it did use were
simply there to apply global parameter values to voltage sources.

It was failing because of the following:
(1) it was doing .STEP sweeps on the global_params used by 2 voltage
sources.
(2) on certain .STEP iterations, it couldn't converge the DCOP unless it
used sourcestepping  (the 3rd option in the "spice strategy" in the
nonlinear solver).
(3) source stepping, in combination with Vsrc's that relied on
expressions, which in turn depended on .STEP-determined global
parameter values, didin't play well together
(4) they didn't play well together because with the old
expression library, the logic depended a great deal upon the "changed"
logic, and I had to disable that logic (temporarily, I hope) with
new expression.

I previously dealt with something very similar, but less complicated.
Earlier in my work, I had the problem of making a global_param-dependent
Vsrc and a .DC sweep on that Vsrc work together.  This was simpler than
the sourcesweep example, because I just had to make sure that the DC
sweep value would override the global_param value, and once it was
overridden, it didn't matter what the global_param value even was.

In the sourcestepping case, it is necessary to make sure that the
global_param value is right, and that it is applied to expressions that
depend on it, and the expression result applied to the relevant device
parameter.  All this must happen BEFORE the sourcestepping iterations
rescale the value of the Vsrc's.    And then AFTER this rescaling
happens, the expression must not be re-evaluated again, until
the global_param is iterated again in the next .STEP iteration.

To make this work, I track parameters that are set via
DeviceMgr::setParam and/or scaleParam, and I put them in a map.  Then,
when updateDependentParameters is called, anything in this map is not
updated.

Also, after the source-scale algorithm completes, the map is cleared of
any "scaled" params.  That way, new global_param values can still be a
pplied.   This is accomplished via DeviceMgr::resetScaleParams, which is
called from the LOCA group.

Also, the "DeviceMgr::updateDepenentParams" function is now a  public
function, and it is called from the "updateSweepParams" function.  That
way, any .STEP or .DC values are applied (even if expression dependent)
BEFORE the sourcestep scaling value can be applied.

The test suite is down to 4 meaningful test failures at this point,
using the follwing tags:

--taglist="+nightly+serial-measure-b4soi-ylin?klu?rad?qaspr-bug812son-commandline "

The high-priority tests, that I thought had a good chance of getting
broken thru this work, correspond to these tags:

--taglist=" +bug1801+hicum_l0+bug1035son+sourcestep?klu "

These all pass.  (but all of them were broken at various points during
this work)