Band-aid Xyce/ADMS for one special case
ADMS itself knows how to handle only "flow contributions" of the form
"I(a,b) <+ (expression)" (denoting that a specified current should
flow between nodes a and b) , and only really knows what to do with
"potential probe" (accessors returning a potential drop V(a,b) or
Temp(node) for use in expressions).

ADMS parses "flow probes" (accessors returning the current flowing
between nodes a and b for use in expressions) and "potential
contributions" (e.g. "V(a,b)<+expression" saying that there should be
a specified potential drop between two nodes") but doesn't actually
know what to do with them.

Back in 2011 in service of issue 1779 on morannon, I hacked in some
support for potential contributions and flow probes (see the issue for
a link to the resource I used to figure out what special cases I
needed to support), and this was sufficient to get some simple use
cases working (inductors, voltage sources, serial RLC toy devices, and
also FBH HBT with parasitic inductors and the MVS 2.0.0 model).

However, there were some bugs.  I got all the uses of flow probes
wrong if there are no potential contributions to go with them, and got
one conditional wrong that incorrectly handled the case where a given
branch (a,b) had both flow and potential contributions.

This latter bug (which is gitlab issue 50) showed up for the first
time when I tried to create a parallel RLC device, and I incorrectly
diagnosed it as an issue with "named branches" and documented it in
the Xyce/ADMS users guide as if that were the case.  I soon realized
that the bug had nothing to do with named branches, but didn't fix it.

This commit fixes that one little bug --- a simple mistake in a
conditional in xyceBasicTemplates.

Now a parallel RLC device without the workaround does in fact work as
expected and as it always should have.  I am working on another CMC
device that hit the same problem, which is why I'm actually getting
around to patching the bug now.

The cases where flow probes are used and there is no corresponding
potential contribution to go with it are all broken, though.  These
things actually behave very differently.  If there are neither flow
nor potential contributions to go with the flow probe usage, then it
is supposed to set up a problem as if there were a zero-volt voltage
source between the nodes -- there are currently missing bits of F
vector load in this case.  If there are flow
contributions and no  potential contributions to a branch, then the
way the contributions are handled has to be completely different, and
this case is completely mishandled, loading all the wrong stuff.

Xyce/ADMS will now emit a warning if it detects that it needs to do
one of the broken special cases.  Turns out that L_UTSOI does in fact
use this in its noise model, but somehow it doesn't break the CMC test
suite.  Ick.  I guess I have to fix that, too.

The entire handling of flow probes and potential contributions needs a
refactor, as it is written in a manner consistent with the state of
Xyce/ADMS in 2011 and is more complex than it should have to be.  This
is the subject of gitlab issue 48.

https://morannon.sandia.gov/bugzilla/show_bug.cgi?id=1779
https://gitlab-ex.sandia.gov/Xyce/Xyce/-/issues/50
https://gitlab-ex.sandia.gov/Xyce/Xyce/-/issues/48