Fix for bug 794 son.
Bug 794 is entitled:
"Bug in TABLE Form of Xyce B-Source (Maybe Caused by a Bug in the Expression Library)"

Tom and Pete dug into this issue a while ago, and the issue ultimately
came down to how derivatives of tables were computed in the old
expression library.

In the new expression library, I started out doing derivatives
completely differently than the old library, via the interpolator
classes.  However, my doing so prevented one of the coilgun tests (the
COILGUN2/coilgun_4-coil_ddx.cir test) from passing.  It would crash part
way thru the simulation using what I though was a reasonable derivative
calculation.   Eventually, out of desperation, I re-implemented the old
expression library's method for computing table derivatives and then it
was able to run the coilgun_4-coil_ddx.cir test and pass.

I should note that the coilgun_4-coil_ddx.cir fails to run all the way
thru using ANY of the interpolators I created in N_UTL_Interpolators.h,
if I am using the "evalDeriv" function from those interpolators.  So, I
think the coilgun_4-coil_ddx.cir is a very problematic circuit.

Anyway, the 794 bug is basically about a very simple 2-entry table
circuit failing, and getting time step too small.  The original circuit
is:
  * test the TABLE form of the G-Source
  .TRAN 0 1us
  .PRINT TRAN V(node1) V(node2)

  R1         node2 0  1
  G1         node2 0 TABLE V(node1,0) = (-2,-3) (2,3)
  V1         node1 0 SIN(0V 0.75V 1e6 0 0 0)

It is basically trivial, and should have no problem running.  But it
doesn't run with the old expression library.   And, as it turns out,
because I duplicated the derivative table calculation, the new
expression library couldn't do it either.

But, it was a reasonably easy fix.    The derivative calculation from
the old expression library would compute derivatives at the midpoints of
each cell of the table, and then do an interpolation on those.  It
did not handle the endpoints very cleanly, and when the table is NOTHING
but endpoints, I think it basically didn't work.  So I've hacked in some
extra code to handle the 2-entry case.

Note that this new hacked 2-entry code does result in 3 regression tests
that had been passing, failing again.  They are:

C_T/BUG_746_SON/raw_file_b_e_h_sources............FAILED[sh]
C_T/BUG_746_SON/raw_file_f_sources................FAILED[sh]
C_T/BUG_746_SON/raw_file_g_sources................FAILED[sh]

All 3 of these tests include a 2-entry table and are failing with minor
differences.  So I am not worried about these.  (ironically, I *just*
got these working recently; they had been failing due to the sqrt of a
negative number)

Also, in this commit; while I was experimenting on this bug I decided to
try out all the interpolators in the N_UTL_Interpolators.h.  I set this
file up a long time ago, and some of the methods in it are used quite a
bit in rad models.  So I think this code is solid.

I discovered that *some* of the interpolators had not been tried with
the new expression library before, and as such, while they are all
templated, some of them didn't play well with complex numbers.  So, this
commit includes fixes to some of the interpolators for this issue.

Then finally, I thought a long time about whether or not the TABLE
operator in the expression library should be restricted to PWL
approximations, especially given that we have these other nice
interpolators available.  Ultimately, for backward compatibility
purposes, I chose to leave TABLE, and the related AST class tableOp
"as-is".

Instead, I have created a new AST operator class, called
"interpolatorOp".  So far it is very similar to the tableOp class,
except that the interpolatorOp::dx function unconditionally calls the
"::evalDeriv" function and does not do the midpoint method.  My
intention is to use this new operator to support other types of
interpolation in the expression library, but connect to new keywords.
So I hope to support tokens such as "SPLINE" and "AKIMA" and "BLI" (for
Barycentric Lagrange Interpolation).    All of these new tokens will
call this operator, and depending on which one calls it, will optionally
allocate the most appropriate interpolator class.  This is new class
still a work in progress, but it compiles.  I haven't yet added new
keywords to the lexer/parser.