Remove workaround in rlc3.va for testing
rlc3.va was written to work around a bug in Xyce/ADMS that was fixed in
gitlab-ex issue #50.

This commit removes the workaround and provides a second netlist that
uses only native Xyce discrete parts instead of the verilog-a module.

rlc_parallel.cir uses the rlc3.va verilog, rlc_parallel_discrete.cir
uses the discrete r, l, and c devices.  The two netlists produce the
same results.

Prior to the fix of #50 Xyce would get convergence problems on this
version of rlc3.va due to the bug in Xyce/ADMS causing broken code to be
emitted.

https://gitlab-ex.sandia.gov/Xyce/Xyce/-/issues/50