More work on improving state management in funcOp
This work is driven by work on the gentleBridgewire.cir test, which is
failing.  That test has many nested functions, and some of those
functions use the SDT (time integral) operator.

The funcOp needs to be careful with internal state, if it includes any
operators in the tree that rely on internal state.

I figured out the "state" issues in a previous commit, at least for the
funcOp::val() function.  Turns out I didn't do it for the funcOp::dx()
function.

This commit partially fixes the dx function, but probably not
completely.    It addresses the first part of the chain rule.