Checkpoint commit,another attempt at SDT/DDT state
My previous attempts at managing the proper state of the SDT and DDT
operators had flaws in them.

This commit contains my best attempt at it yet.  There is still a
problem with it, and this is still a work in progress.  But it is the
closest I've gotten so far.

The main use case that I've been trying to figure out is what to do when
you have a tree of .funcs, and some of them include SDT operators, and
the SDT operators are operating on the arguments to the the funcs.  Oh,
and some of the .funcs are called in more than one place, meaning that a
single unique SDT operator class has to compute integrals for (potentially)
more than one unique argument.  So if you have this:

.func f1(a) {sdt(a)}
.func f2(b) {f1(b)}
{f2(v(1)) + f2(v(3))}

You should have the equivalent of:   {sdt(v(1)) + sdt(v(3))}

But, in the .func tree example, there is a single unique AST class
for f1, and a single  unique AST class for sdt.  And, f1 by itself, of
course doesn't know anything about how it is called.  It doesn't
know that it is called from f2, and f2 doesn't know that it
gets called 2x in the expression, but using different args
(v(1) and v(3)).

But when the .func-ized expression is evaluated, each .func call does a
"setNode" call on the function argument pararms.  So, essentially, in
the final expression, you have funcOp class representing f2, and it has
an internal node that has been set to point at {f1(b)}.  When it is
called, the b parameter also has a "setNode" to first attach v(1) and
then attach v(3).  But then, the f2 expression contains a funcOp class
that has an internal node pointing at {sdt(a)}, and the "a" parameter
must have a setNode performed as well, to point at the "b" paraemter in
f2, which points at the v(1) and/or v(3).  etc.

So this is a bit confusing.  I didn't write the function handling (or
any of the expression library) to pass internal state around, just
arguments to operators.

After trying several other ideas, I gave in to an idea I had been resisting.
I ultimately decided that I needed to included sdt/ddt state information
in every operator class.  At the moment, this state information is
uncondtionally allocated, but later on I think this can be
improved so that it is only allocated when needed.

Since the state is contained in the operators, then all the operator
propgation management that I was already doing for function arguments
should automatically carry the correct "state" information with it.  At
least, that is the theory.  So, if the function argument propagation
works correctly (and I think it does, or we're in bigger trouble), then
the accompanying state information should also propagate correctly.

I think I've done most of the work to get this set up correctly.
However, the .func argument handling was tricky, as the "state" that
needs to be tracked is a couple of degrees removed from what the SDT
operator sees when called via .func.

In the course of doing this, I did a number of things.  There is now
some static data in the ast base class, and to initialize it correctly
there is now an ast.C file that exists only for this initialization.A

One thing that the static data is doing is managing unique IDs for every
unique AST class instance.  Since I don't do string substitution, and I
don't really copy AST classes ever either (I just copy the RCPs),  it is
easy to accidentally perform an operation 2x on the same operator,
depending on the structure of the AST tree.  Having a unique ID makes it
easier to avoid this.  This was particularly important for managing
the history of SDT and DDT.

Most of the other static bits I've added are also related to managing
the SDT and DDT history.   This is still a work in progress and may
change.

One caveat about all of this:  Xyce with all this new state management
still doesn't pass the "gentleBridgewire" test.  But, the result it
producces looks 100% identical to the gold standard, just shifted a bit,
and requiring more time steps.  I remember observing this when I worked
on SDT in the old expression library, and this may just be a symptom of
that circuit being really hard to solve.  (it has a massive feedback
loop in the Bsrc, which could cause problems).  So, even though it
doesn't pass, I may eventually prove to myself that this is OK.

Finally, for this commit, I am leaving in place a LOT of commented-out
debug output.  In my recent work I learned the hard way not to remove
that stuff too hastily.  This code compiles and (seems to) work.  So, I
want to save my work before hacking and slashing away at the all the
commented-out dead code.