Add more unit tests for SDT with lots of funcs
I believe my recent design for state management for SDT/DDT is
the correct one, other than possible mistakes.  So, I am now attempting
to conceive of unit tests that will break it.

This commit adds 3 new tests that involve various levels of .func
nestinng with SDT operators.  So far, so good.  The deepest tree is 100
funcs deep and it gets the right answer.  I've also experimented a bit
with a smaller tree, but with different types of sdt calls at each
layer.  Also, so far, so good.