Refine SDT/DDT state management for SDT of a func
Most of the use cases I've tested have been for funcs of SDT operators.
In other words:

.func f2(a) {sdt(a)}
.func f1(a) {f2(a)}
{f1(a)}

But, I hadn't tested (at least in unit testing) SDTs of funcs.  ie:

.func f2(a) {cos(a)}
.func f1(a) {f2(a)}
{sdt(f1(a))}

This commit adds a few tests of this nature.  It also adds some
virtual functions to the funcOp class (which are relevant to these
tests) that I had neglected to add.  My previous work was mostly
focused on the paramOp class, and these virtual functions have
the same structure in funcOp as they did in paramOp.