Another state-management fix for SDT.
My previous commits on this were a step in the correct direction, but
not 100% correct.   This commit is an improvement, but also still not
100% correct.  More on that later.

In my previous commits, to be expeditious I chose to rely on some static
data to avoid processing the state data associated with an SDT operator
more than once.  This was a concern for circuits in which many
expressions referred to the same .func, or .funcs that used an SDT
operator.  All of those func calls would ultimately interact with the
same, unique sdtOp class.  And, if they used the same argument to SDT,
then they potentially also interacted with the same argument operator
multiple times.

Relying on static data in the ast.h file was fine (although I generally
dislike doing that), but my management of the static data had (at least)
one mistake.  One of the state objects is a map object, which tracked
which arguments to the sdt operator had already rotated their state when
the time step advanced.  That design was fine, I think, but I was
clearing that map in the newExpression::processSuccessfulTimeStep
function, and that was not fine.    If two different newExpression
objects ultimately referred to the same sdtOp class, and used the same
arguments to sdtOp, then clearing this map would result in it being
operated on 2x (or more).  That caused the gentleBridgewire test to
exibit some weird behavior.

With this commit, I have moved that "clear" function out of the
newExpression::processSuccessfulTimeStep function, and put it into a
static function that must be called once during the processing of
successful time steps.  It is currently called from the
DeviceMgr::acceptStep function.  As it operates on static data, it made
sense to have it be updated via a static function.

As of this commit, the bridgewire test now passes and behaves ~98%
as it should.  Technically, if the regression test suite is all we have
to go on, that would be enough.  But I happen to know that
it is *still* not really quite correct.

This commit takes care of most of the issues, but there is at
least one issue I have identifified that still needs fixing.
Also, even to the extent that this implementation correct, the
state handling that I have set up as of this commit still
needs cleanup; what is in this commit was really more of a
"proof-of-concept" for an idea that I had.

So, there will be more commits to come, hopefully today.