Created & applied new function, createFunc
Creating and fully resolving a .func, in the manner done by Xyce, takes
a number of steps, which leads to roughly 10 lines of code.  Much of
those 10 lines are always the same and include temporary data structures
that aren't needed once resolution is over.  This led to pretty messy
code whenever I was using a lot of functions.

So, I've created a simplyfying function, called createFunc.  So far I've
applied it to 2 tests.