More cleanup, application of createFunc
The "createFunc" function was added in a previous commit.  The point of
it is to simply the creation and resolution of .func objects in unit
tests.  There were many tests that had .funcs in them, so swapping them
out for the new function is a slow process.

In addition to that work, quite a few sdt and ddt tests needed to
be updated to test the copy constructors and assignment operators.
When I am first developing a test I usually leave those tests
out and then add them later once the test is mature.  But, it is
easy to forget them later.  So this is me finally getting
around to that.

Also, part of this commmit includes more correct usage of
"processSuccessfulTimeStep", which is a fragile aspect of SDT and DDT.