Pass Final Sim Time into TRAN Measure Objects
This commit changes the function signatures of the various
updateTran() functions of the measure classes, and the tranOutput()
function in the OutputMgrAdapter class, to pass the "final simulation
time" to the Measure classes.  This additional information will be
useful for the processing the AVG, DERIV-WHEN, FIND-WHEN, INTEG, RMS
and WHEN measures, as I work on making the FROM-TO and RISE/FALL/CROSS
qualifiers work better for those measure types.

This commit has no immediate impact on the behavior of those
measure types though, and has no impact on the behavior of the
outputter classes.

The approach taken (of allocating the analysis object for the
RemeasureDC and remeasureTRAN classes will not work for passing
the start and end simulation frequencies into the RemeasureAC object,
because the AC object constructor segfaults on the initialization
of its bVecRealPtr and bVecImgPtr member variables when that
initialization is done from within the RemeasureAC constructor,
rather than via its "normal path".  Since -remeasure is not currently
supported for AC measures, for other reasons, this was deemed okay
for now.  The existing tests for -remeasure for AC measures
(MEASURE_AC/ErrorTestAC and MEASURE_AC/STEP/ErrorTestAC) still pass
since the ERROR measure does not support the FROM and TO qualifiers.
The Bugzilla entry for SON Bug 1141 will discuss other approaches
for making -remeasure work better for .AC.

SON Bugs 1141, 1304, 1309 and 1310
http://joseki-vm.sandia.gov/bugzilla/show_bug.cgi?id=1141
http://joseki-vm.sandia.gov/bugzilla/show_bug.cgi?id=1304
http://joseki-vm.sandia.gov/bugzilla/show_bug.cgi?id=1309
http://joseki-vm.sandia.gov/bugzilla/show_bug.cgi?id=1310