Add unit test that demonstrates sdt state failure
The sdt state management that I recently set up *almost* does everything
it needs to, but not quite.

In this commit, there are two tests, sdt15 and sdt16, which demonstrate
a remaining problem.  The sdt15 test passes.  The sdt16 test does not.

sdt16 is very similar to sdg15.  Both of them test what happens when
multiple expressions call the same .func which happens to include an sdt
operator.  What is different is the argument to the sdt operator.

In the sdt15 test, the .func is F2(B) {sdt(B)}.  In other words, exactly
the parameter that is passed in, is the subject of the sdt operation.

In the sdt16 test, the .func is F2(B) {sdt(B*B)}.  In other words, the
sdt operator contains an expresion which uses B, rather than just B.
So, the the AST node seen by the sdt operator in this case will be the
binaryMulOp (binary multiply).  It doesn't "see" the B parameter
directly, because the B parameter is not at the top of the AST tree.

My design mistake has been to assume that the "state" information needed
to track with parameter args (and other things) like B.  But that isn't
correctly.  The state information needs to track with whatever is the
argument to sdt.