Final fix for sdt/ddt state management
I *think* I have finally got it right.  This was tricky.

The short version of this is that I needed to ensure that any SDT or DDT
operator had the correct state information.    The following scenarios
needed to be handled:

(1) when sdt/ddt was inside of a .func
(2) when that .func was called at the end of a chain of .func calls
(3) be correct even when the argument to sdt/ddt was another expression
(4) be correct even if multiple expressions were calling the same .func,
using completely different (or the same) arguments.A
(5) recognize when the same state should be used, and when it should
not. (this has a few complexities)

A small, but complicated case, which is only fixed as of
this commit, is this.  What I am describing is the unit test:

TEST(Double_Parser_Integral_Test, sdt16)

.func F1(A) {f2(A)}
.func F2(B) {sdt(B*B)}
expr1 = {f1(V(A))}
expr2 = {f1(V(B))}

.func F3(A) {f4(A)}
.func F4(B) {sdt(B*B)}
expr3 = {f3(V(A))}

V(A) is set to be a non-trivial time-dependent function (like
sine wave), and V(B) is set to some other completely different
non-trivial time-dependent funciton (like V(B) = time).

expr3 is identical to expr1, except that it calls a set of functions
that are NOT called by any other expression.   So, there are no concerns
about it being corrupted by another expression.  expr1 can be corrupted
by expr2 (if handled incorrectly) but expr3 can't be corrupted by
anyone.

The goal of the test is for the result of expr1->val() to identically
match the result of expr3->val().  And not just match, but be identical.
This is a concern b/c it is easy for expr1 to be corrupted by expr2.

As of this commit, this test passes. All the other problematic
sdt cases that I was already testing still pass, also.
So, this work is about done.

I'm not 100% happy with the code structure in this commit, but
some of that was simply to temporarily avoid doing more typing
than I had to while testing out ideas.  I will clean it up in
subsequent commits.

For example, I rotate the time step data inside of a val() call.  It was
convenient to do this b/c the call structure traversing the whole AST
tree was already in place, and I could be confident that the function
arguments that needed to be set properly would be during the traversal.
But it really should have its own set of functions.

Also, there are some (mostly harmless) artifacts of other ideas I tried
to solve this.  They should go.