Added another sdt test, with mutliple func args
Another complexity in sdt state management is ensuring that unique
states are used for sdt's of expressions, when the expression might
use all the same function arguments, but applied in a different order.

In more detail, if you have:

.func f1(a,b) {f2(a,b)}
.func f2(a,b) {sdt(a+b*b)}

expr1 = {f1(V(A),V(B)}
expr2 = {f1(V(B),V(A)}

Then, expr1 and expr2 should obviously produce different answers.
However, from the point of view of the sdt operator, whenever it is
called, it only needs a single state object to manage its history.

expr1 and expr2 call the same functions, so they are using the same AST
node classes for evaluation.  So, done really naively, the sdt operator
would have a single state object that was owned locally and attempt to
use it and get the wrong answer (for both as it each would corrupt the
other)

A slightly less naive thing to do is to check the arguments to sdt, and
then have a unique state for each unique set of arguments.  But, this
has to be done carefully, as the order of the arguments matter as well.
ie, f2(a,b) should use a different state object than f2(b,a).

Also, if you have multiple expressions calling f2(a,b) (same arguments,
same order), you have to make sure that you only advance the state once,
even if 100 expressions all correctly use that state.

This new test is the first of several to attempt to test out the
2-argument issue(s).